Partially Address Comments ## No Longer Destructuring Messages Array I set up mapStateToProps within the messages component so it renders whenever state is updated. This removes the O(N) time complexity for each render.
## Update Read Status in DB on Backend
All messages matching the active/selected conversationId are now marked
as read when the API call is made, as opposed to running conditionally
with every messages on every render.

## Unread Message Count Moved to Backend
When conversations are first fetched and converted to JSON, they now
have their unread messages counted in an array. That count is then
accessible through state in the chat content component to be more
efficient.

## Message Status Update on Active Chat Broken
The reason for destructuring the array in my last iteration was that
small updates like changing the read property of a single message
for a specific conversation was not rerending the Messages component.
Removing the spread was necessary for efficiency, but has removed
functionality; this will be fixed in next commit.